Move twig tag parsers into seperate module
